renamed a file; moved graph processing to graph util; created schematics config utils
